Update patch set 1
Patch Set 1:

this is a global property file, so ifs it not global it shouldn't be here. Also test case specific variables should be in the test case folder. I did have a conversation with yuriy around allowing the servicemapping structure have the ability to accommodate cds data like it does preload but that is not ready yet.

Patch-set: 1