Update patch set 9
Patch Set 9:

What is the flag to set so that we can specifically test either the ingress port or the internal k8 port for the service. As a test tool I think we need to be able to confirm that internally components can talk to each other and separate test that the ingress controller path is working for those applications that also have an ingress controller port.

Patch-set: 9
Reviewer: Gerrit User 29 <29@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>