Update patch set 2
Patch Set 2: Code-Review-1

In my humble opinion it's a terrible idea.

As far as I understand (let me know if I'm wrong) this would make robot depend on AAF and we don't want that.

Robot is not considered to be integral part of ONAP. At some point we may want to deploy robot totally outside of k8s cluster in which ONAP is running to make the tests more from the end user perspective. In my opinion if you need a certificate you should retrieve it as a part of your testcase and clearly fail the test if AAF is not responding but don't try to integrate robot pod with AAF

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1