Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Do we need to remove healthcheck from the old legacy engine?? Is that still there? I thought there was something.

I removed the legacy healthchecks in the OOM review (refer to removal of the custom feature-healtcheck.properties custom configuration that enabled it), so this does not exercise it the legacy policy engine healthchecks) and should pass healthchecks in combination with the other OOM review as is.

We could remove the drools-pdp healthcheck altogether, if we decide so, but I suggest we do it in another review.   I would consider it a different work item.   I would prefer not to add more complexity to the OOM review plus this necessary companion one that should pass the healtcheck, as it is a pretty large piece of work as is.    It will be easier to add any additional enhancements that we consider necessary later on top of it after these 2 are merged IMO.

Patch-set: 2